GraphControl skips Infinite/NaN values. Enhanced PlotCurveListView functionality. SpecControl data management and plotting. Updated DummyAndorCamera yet again. Removed redundant cast from Data.DeltaOD(double[], double[]).
Closes #39.

+++ lasercom/Data.cs
@@ -204,7 +204,7 @@ namespace lasercom
         {
             double[] OD = new double[Ground.Count];
             for (int i = 0; i < OD.Length; i++)
-                OD[i] = Math.Log10((double)(Trans[i] - Dark[i]) / (double)(Ground[i] - Dark[i]));
+                OD[i] = Math.Log10((Trans[i] - Dark[i]) / (Ground[i] - Dark[i]));
             return OD;
         }
